archive/requirements/base.txt
# FrickTrader Base Requirements
# Core dependencies required in all environments

# Core Framework
freqtrade==2024.9

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Technical Analysis
ta-lib>=0.4.24
pandas-ta>=0.3.14b0

# Exchange APIs
ccxt>=4.4.98
python-binance>=1.0.17

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# WebSockets
websockets>=11.0.0

# Caching
cachetools>=5.3.0

# CLI
click>=8.1.0
rich>=13.0.0

# Async Support
asyncio-mqtt>=0.13.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3